+++ packages/web3-eth/test/integration/subscription_heads.test.ts
@@ -43,7 +43,27 @@ describeIf(isSocket)('subscription', () => {
 			let times = 0;
 			const pr = new Promise((resolve: Resolve, reject) => {
 				sub.on('data', (data: BlockHeaderOutput) => {
-					expect(typeof data.parentHash).toBe('string');
+					try {
+						expect(typeof data.hash).toBe('string');
+						expect(typeof data.parentHash).toBe('string');
+						expect(typeof data.receiptsRoot).toBe('string');
+						expect(typeof data.miner).toBe('string');
+						expect(typeof data.stateRoot).toBe('string');
+						expect(typeof data.transactionsRoot).toBe('string');
+						expect(typeof data.logsBloom).toBe('string');
+						expect(typeof data.difficulty).toBe('bigint');
+						expect(typeof data.number).toBe('bigint');
+						expect(typeof data.gasLimit).toBe('bigint');
+						expect(typeof data.gasUsed).toBe('bigint');
+						expect(typeof data.timestamp).toBe('bigint');
+						expect(typeof data.extraData).toBe('string');
+						expect(typeof data.nonce).toBe('bigint');
+						expect(typeof data.sha3Uncles).toBe('string');
+						expect(typeof data.baseFeePerGas).toBe('bigint');
+						expect(typeof data.mixHash).toBe('string');
+					} catch (error) {
+						reject(error);
+					}
 
 					times += 1;
 					expect(times).toBeGreaterThanOrEqual(times);