+++ packages/web3-types/src/eth_types.ts
@@ -144,18 +144,42 @@ export interface BlockOutput {
 	readonly parentHash?: HexString32Bytes;
 }
 
+export interface Withdrawals {
+	readonly index: Numbers;
+	readonly validatorIndex: Numbers;
+	readonly address: Address;
+	readonly amount: Numbers;
+}
+
 export interface BlockHeaderOutput {
+	readonly hash?: HexString32Bytes;
+	readonly parentHash?: HexString32Bytes;
+	readonly receiptsRoot?: HexString32Bytes;
+	readonly miner?: HexString;
+	readonly stateRoot?: HexString32Bytes;
+	readonly transactionsRoot?: HexString32Bytes;
+	readonly withdrawalsRoot?: HexString32Bytes;
+	readonly logsBloom?: Bytes;
+	readonly difficulty?: Numbers;
+	readonly number?: Numbers;
 	readonly gasLimit: Numbers;
 	readonly gasUsed: Numbers;
 	readonly timestamp: Numbers;
-	readonly number?: Numbers;
-	readonly difficulty?: Numbers;
-	readonly totalDifficulty?: Numbers;
-	readonly transactions?: TransactionOutput[];
-	readonly miner?: HexString;
-	readonly baseFeePerGas?: Numbers;
-	readonly parentHash?: HexString32Bytes;
+	readonly extraData?: Bytes;
+	readonly nonce?: Numbers;
 	readonly sha3Uncles: HexString32Bytes[];
+	readonly baseFeePerGas?: Numbers;
+
+	// These fields are returned when the RPC client is Nethermind,
+	// but aren't available in other clients such as Geth
+	readonly author?: Address;
+	readonly totalDifficulty?: Numbers;
+	readonly size?: Numbers;
+	readonly excessDataGas?: Numbers;
+	readonly mixHash?: HexString32Bytes;
+	readonly transactions?: TransactionOutput[];
+	readonly uncles?: Uncles;
+	readonly withdrawals?: Withdrawals[];
 }
 
 export interface ReceiptInput {
@@ -471,4 +495,18 @@ export interface AccountObject {
 	readonly storageHash: Bytes;
 	readonly accountProof: Bytes[];
 	readonly storageProof: StorageProof[];
+}
+
+export interface Eip712TypeDetails {
+	name: string;
+	type: string;
+}
+export interface Eip712TypedData {
+	readonly types: {
+		EIP712Domain: Eip712TypeDetails[];
+		[key: string]: Eip712TypeDetails[];
+	};
+	readonly primaryType: string;
+	readonly domain: Record<string, string | number>;
+	readonly message: Record<string, unknown>;
 }